Changed intervals and constants to make game spawn more enemies.

+++ examples/game/statelabel.cpp
@@ -65,7 +65,7 @@ void StateLabel::updateGameStatus()
 					  << m_game->player()->hitPoints()
 					  << m_game->player()->score()
 					  << m_game->highscore()
-					  << sad::Renderer::ref()->fps()
+					  << (int)(sad::Renderer::ref()->fps())
   );
   string() = s;
 }
